requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.1.0
    # via quart
aiohttp==3.8.4
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.5
    # via quart
boto3==1.26.149
    # via -r requirements.in
botocore==1.29.149
    # via
    #   boto3
    #   s3transfer
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via quart
cryptg==0.4.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.14.3
    # via
    #   -r requirements.in
    #   quart
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via quart
jinja2==3.1.2
    # via quart
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.2
    # via
    #   jinja2
    #   quart
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.1
    # via pytest
pluggy==1.0.0
    # via pytest
priority==2.0.0
    # via hypercorn
pyaes==1.6.1
    # via pyrogram
pydantic==1.10.7
    # via -r requirements.in
pyrogram==2.0.106
    # via -r requirements.in
pysocks==1.7.1
    # via pyrogram
pytest==7.3.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
python-dotenv==1.0.0
    # via -r requirements.in
quart==0.18.4
    # via -r requirements.in
requests==2.30.0
    # via -r requirements.in
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
tgcrypto==1.2.5
    # via -r requirements.in
toml==0.10.2
    # via hypercorn
typing-extensions==4.5.0
    # via pydantic
urllib3==1.26.16
    # via
    #   botocore
    #   requests
werkzeug==2.3.4
    # via quart
wsproto==1.2.0
    # via hypercorn
yarl==1.9.2
    # via aiohttp
